Updated Acceptance Tests
- Logging in with LMS before all tests
- Skipping error page tests if the pages are not enabled
