+++ acceptance_tests/test_auth.py
@@ -1,32 +1,27 @@
-from bok_choy.web_app_test import WebAppTest
+from unittest import skipUnless
 
-from acceptance_tests import USERNAME, PASSWORD, ENABLE_OAUTH_AUTHORIZE
+from bok_choy.web_app_test import WebAppTest
+from acceptance_tests import ENABLE_OAUTH_TESTS
+
+from mixins import LoginMixin
 from pages import LoginPage
 
 
-class OAuth2FlowTests(WebAppTest):
+@skipUnless(ENABLE_OAUTH_TESTS, 'OAuth tests are not enabled.')
+class OAuth2FlowTests(LoginMixin, WebAppTest):
     def setUp(self):
         """
         Instantiate the page objects.
         """
         super(OAuth2FlowTests, self).setUp()
 
-        self.login_page = LoginPage(self.browser)
-
-        self.username = USERNAME
-        self.password = PASSWORD
+        self.insights_login_page = LoginPage(self.browser)
 
     def test_login(self):
+        self.login_with_lms()
+
         # Visit login URL and get redirected
-        self.login_page.visit()
-
-        # Input credentials and authorize
-        self.login_page.q(css='input[name=username]').fill(self.username)
-        self.login_page.q(css='input[name=password]').fill(self.password)
-        self.login_page.q(css='button[type=submit]').click()
-
-        if ENABLE_OAUTH_AUTHORIZE:
-            self.login_page.q(css='input[name=authorize]').click()
+        self.insights_login_page.visit()
 
         # User should arrive at course index page (or access denied page, if no permissions)
         # Splitting this out into two separate tests would require two separate sets of credentials. That is
