+++ acceptance_tests/test_course_engagement.py
@@ -3,7 +3,7 @@ import datetime
 from bok_choy.web_app_test import WebAppTest
 
 from analyticsclient.constants import activity_type as at
-from acceptance_tests import CoursePageTestsMixin
+from acceptance_tests.mixins import CoursePageTestsMixin
 from acceptance_tests.pages import CourseEngagementContentPage
