+++ acceptance_tests/test_error_pages.py
@@ -1,8 +1,10 @@
+from unittest import skipUnless
 from bok_choy.web_app_test import WebAppTest
-from acceptance_tests import PLATFORM_NAME, APPLICATION_NAME, SUPPORT_URL
+from acceptance_tests import PLATFORM_NAME, APPLICATION_NAME, SUPPORT_URL, ENABLE_ERROR_PAGE_TESTS
 from pages import ServerErrorPage, NotFoundErrorPage, AccessDeniedErrorPage, AuthErrorPage
 
 
+@skipUnless(ENABLE_ERROR_PAGE_TESTS, 'Error page tests are not enabled.')
 class ErrorPagesTests(WebAppTest):
     error_page_classes = [ServerErrorPage, NotFoundErrorPage, AccessDeniedErrorPage, AuthErrorPage]
